Created apollo plugin to automatically inject the selection map into the context so it is availabe for each request. Fixed the plugin assignement (accidentally wrapped it array). Move selection code to own file.
