+++ opencti-platform/opencti-graphql/src/cyio/schema/assets/software/sparql-query.js
@@ -89,15 +89,28 @@ const predicateMap = {
   }
 }
 
-const selectQueryForm = `
-SELECT ?iri ?id ?object_type 
-  ?asset_id ?name ?description ?locations ?responsible_party 
-  ?asset_type ?asset_tag ?serial_number ?vendor_name ?version ?release_date
-  ?function ?cpe_identifier ?software_identifier ?patch ?installation_id ?license_key
-FROM <tag:stardog:api:context:named>
-WHERE {
-    ?iri a <http://scap.nist.gov/ns/asset-identification#Software> .
-`;
+const predicateBody = `
+    ?iri <http://darklight.ai/ns/common#id> ?id .
+    OPTIONAL { ?iri <http://scap.nist.gov/ns/asset-identification#asset_id> ?asset_id } .
+    OPTIONAL { ?iri <http://scap.nist.gov/ns/asset-identification#name> ?name } .
+    OPTIONAL { ?iri <http://scap.nist.gov/ns/asset-identification#description> ?description } .
+    OPTIONAL { ?iri <http://scap.nist.gov/ns/asset-identification#locations> ?locations } .
+    # OPTIONAL { ?iri <http://scap.nist.gov/ns/asset-identification#responsible_parties> ?responsible_party } .
+    # ItAsset
+    OPTIONAL { ?iri <http://scap.nist.gov/ns/asset-identification#asset_type> ?asset_type } .
+    OPTIONAL { ?iri <http://scap.nist.gov/ns/asset-identification#asset_tag> ?asset_tag } .
+    OPTIONAL { ?iri <http://scap.nist.gov/ns/asset-identification#serial_number> ?serial_number } .
+    OPTIONAL { ?iri <http://scap.nist.gov/ns/asset-identification#vendor_name> ?vendor_name }.
+    OPTIONAL { ?iri <http://scap.nist.gov/ns/asset-identification#version> ?version } .
+    OPTIONAL { ?iri <http://scap.nist.gov/ns/asset-identification#release_date> ?release_date } .
+    # Software - OperatingSystem - ApplicationSoftware
+    OPTIONAL { ?iri <http://scap.nist.gov/ns/asset-identification#function> ?function } .
+    OPTIONAL { ?iri <http://scap.nist.gov/ns/asset-identification#cpe_identifier> ?cpe_identifier } .
+    OPTIONAL { ?iri <http://scap.nist.gov/ns/asset-identification#software_identifier> ?software_identifier } .
+    OPTIONAL { ?iri <http://scap.nist.gov/ns/asset-identification#patch_level> ?patch } .
+    OPTIONAL { ?iri <http://scap.nist.gov/ns/asset-identification#installation_id> ?installation_id }
+    OPTIONAL { ?iri <http://scap.nist.gov/ns/asset-identification#license_key> ?license_key } .
+    `;
 
 const bindIRIClause = `\tBIND(<{iri}> AS ?iri)\n`;
 const typeConstraint = `?iri a <http://scap.nist.gov/ns/asset-identification#{softwareType}> . \n`;
@@ -166,8 +179,7 @@ export const removeFromInventoryQuery = (id) => {
 export const updateSoftwareQuery = (id, input) => {
   const iri = `<http://scap.nist.gov/ns/asset-identification#Software-${id}>`;
   let deletePredicates = [], insertPredicates = [], replaceBindingPredicates = [];
-  for(const change of input) {
-    const {key, value, operation} = change;
+  for(const {key, value, operation} of input) {
     if(!predicateMap.hasOwnProperty(key)) continue;
     for(const itr of value) {
       const predicate = predicateMap[key].binding(iri, itr);
@@ -225,9 +237,9 @@ export const QueryMode = {
 
 export function getSelectSparqlQuery(type, select, id, filter) {
   let sparqlQuery;
-  const { clause, predicates } = buildSelectVariables(predicateMap, select)
+  const { selectionClause, predicates } = buildSelectVariables(predicateMap, select)
   const selectPortion = `
-SELECT DISTINCT ${clause}
+SELECT DISTINCT ${selectionClause}
 FROM <tag:stardog:api:context:named>
 WHERE {
   `;
